Move useShouldNavigateAway to NavigationHandler.kt
